requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-generativeai>=0.8.0
PyMuPDF>=1.24.0
aiofiles>=24.1.0
chromadb>=0.4.22
sentence-transformers>=2.5.0
